refactor: RestDocs를 위한 테스트코드 제거
