+++ src/main/java/wooteco/subway/line/dto/LineRequest.java
@@ -1,7 +1,5 @@
 package wooteco.subway.line.dto;
 
-import javax.validation.constraints.Email;
-import javax.validation.constraints.Max;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
