--- src/test/java/wooteco/subway/auth/AuthAcceptanceRestDoc.java
@@ -1,118 +1,0 @@
-package wooteco.subway.auth;
-
-import io.restassured.response.ExtractableResponse;
-import io.restassured.response.Response;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
-import wooteco.subway.RestDocs;
-import wooteco.subway.auth.dto.TokenResponse;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import static wooteco.subway.DocsIdentifier.*;
-import static wooteco.subway.auth.AuthAcceptanceTest.회원_등록되어_있음;
-import static wooteco.subway.member.MemberAcceptanceTest.회원_정보_조회됨;
-
-public class AuthAcceptanceRestDoc extends RestDocs {
-    private static final String EMAIL = "email@email.com";
-    private static final String PASSWORD = "password";
-    private static final Integer AGE = 20;
-
-    public static TokenResponse 로그인되어_있음(String email, String password) {
-        ExtractableResponse<Response> response = 로그인_요청(email, password);
-        return response.as(TokenResponse.class);
-    }
-
-    public static ExtractableResponse<Response> 로그인_요청(String email, String password) {
-        Map<String, String> params = new HashMap<>();
-        params.put("email", email);
-        params.put("password", password);
-
-        return given(AUTH_LOGIN_SUCCESS).
-                contentType(MediaType.APPLICATION_JSON_VALUE).
-                body(params).
-                when().
-                post("/login/token").
-                then().
-                log().all().
-                statusCode(HttpStatus.OK.value()).
-                extract();
-    }
-
-    public static ExtractableResponse<Response> 내_회원_정보_조회_요청(TokenResponse tokenResponse) {
-        return given(MEMBERS_ME_GET).
-                auth().oauth2(tokenResponse.getAccessToken()).
-                accept(MediaType.APPLICATION_JSON_VALUE).
-                when().
-                get("/members/me").
-                then().
-                log().all().
-                statusCode(HttpStatus.OK.value()).
-                extract();
-    }
-
-    @DisplayName("Bearer Auth")
-    @Test
-    void myInfoWithBearerAuth() {
-        // given
-        회원_등록되어_있음(EMAIL, PASSWORD, AGE);
-        TokenResponse tokenResponse = 로그인되어_있음(EMAIL, PASSWORD);
-
-        // when
-        ExtractableResponse<Response> response = 내_회원_정보_조회_요청(tokenResponse);
-
-        // then
-        회원_정보_조회됨(response, EMAIL, AGE);
-    }
-
-    @DisplayName("Bearer Auth 로그인 실패")
-    @Test
-    void myInfoWithBadBearerAuth() {
-        회원_등록되어_있음(EMAIL, PASSWORD, AGE);
-
-        Map<String, String> params = new HashMap<>();
-        params.put("email", EMAIL + "OTHER");
-        params.put("password", PASSWORD);
-
-        given(AUTH_LOGIN_FAIL)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(params)
-                .when().post("/login/token")
-                .then().log().all()
-                .statusCode(HttpStatus.UNAUTHORIZED.value());
-    }
-
-    @DisplayName("Bearer Auth 유효하지 않은 토큰")
-    @Test
-    void myInfoWithWrongBearerAuth() {
-        TokenResponse tokenResponse = new TokenResponse("accesstoken");
-        given(MEMBERS_ME_GET_WRONG_TOKEN)
-                .auth().oauth2(tokenResponse.getAccessToken())
-                .accept(MediaType.APPLICATION_JSON_VALUE)
-                .when().get("/members/me")
-                .then().log().all()
-                .statusCode(HttpStatus.UNAUTHORIZED.value());
-    }
-
-    @DisplayName("이메일 중복 검사 - 성공")
-    @Test
-    void emailDuplicate_success() throws Exception {
-        given(MEMBERS_ME_GET_CHECK_VALIDATION_SUCCESS)
-                .when().get("/members/check-validation/?email=Other_login%40email.com")
-                .then().log().all()
-                .statusCode(HttpStatus.OK.value());
-    }
-
-    @DisplayName("이메일 중복 검사 - 실패")
-    @Test
-    void emailDuplicate_fail() throws Exception {
-        회원_등록되어_있음(EMAIL, PASSWORD, AGE);
-        given(MEMBERS_ME_GET_CHECK_VALIDATION_FAIL).log().all()
-                .when().get("/members/check-validation/?email=" + EMAIL)
-                .then().log().all()
-                .statusCode(HttpStatus.UNPROCESSABLE_ENTITY.value());
-    }
-}
