+++ src/test/java/wooteco/subway/line/LineAcceptanceTest.java
@@ -1,6 +1,5 @@
 package wooteco.subway.line;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.BeforeEach;
@@ -8,7 +7,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.restdocs.RestDocumentationContextProvider;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.DocsIdentifier;
 import wooteco.subway.line.dto.LineRequest;
 import wooteco.subway.line.dto.LineResponse;
 import wooteco.subway.line.dto.MapResponse;
@@ -19,6 +20,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static wooteco.subway.DocsIdentifier.*;
 import static wooteco.subway.station.StationAcceptanceTest.지하철역_등록되어_있음;
 
 @DisplayName("지하철 노선 관련 기능")
@@ -28,18 +30,20 @@ public class LineAcceptanceTest extends AcceptanceTest {
     private LineRequest lineRequest1;
     private LineRequest lineRequest2;
 
-    public static LineResponse 지하철_노선_등록되어_있음(String name, String color, StationResponse upStation, StationResponse downStation, int distance, Long extraFare) {
-        LineRequest lineRequest = new LineRequest(name, color, upStation.getId(), downStation.getId(), distance, extraFare);
-        return 지하철_노선_등록되어_있음(lineRequest);
+    @BeforeEach
+    public void setUp(RestDocumentationContextProvider restDocumentation) {
+        super.setUp(restDocumentation);
+
+        // given
+        강남역 = 지하철역_등록되어_있음("강남역");
+        광교역 = 지하철역_등록되어_있음("광교역");
+
+        lineRequest1 = new LineRequest("신분당선", "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
+        lineRequest2 = new LineRequest("구신분당선", "bg-red-600", 강남역.getId(), 광교역.getId(), 15);
     }
 
-    public static LineResponse 지하철_노선_등록되어_있음(LineRequest lineRequest) {
-        return 지하철_노선_생성_요청(lineRequest).as(LineResponse.class);
-    }
-
-    public static ExtractableResponse<Response> 지하철_노선_생성_요청(LineRequest params) {
-        return RestAssured
-                .given().log().all()
+    public static ExtractableResponse<Response> 지하철_노선_생성_요청(LineRequest params, String identifier) {
+        return given(identifier)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(params)
                 .when().post("/lines")
@@ -48,27 +52,32 @@ public class LineAcceptanceTest extends AcceptanceTest {
     }
 
     private static ExtractableResponse<Response> 지하철_노선_목록_조회_요청() {
-        return RestAssured
-                .given().log().all()
+        return given(LINES_GET)
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/lines")
                 .then().log().all()
                 .extract();
     }
 
+    private static ExtractableResponse<Response> 지하철_전체_노선도_조회_요청() {
+        return given(LINES_GET_MAP)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/lines/map")
+                .then().log().all()
+                .extract();
+    }
+
     public static ExtractableResponse<Response> 지하철_노선_조회_요청(LineResponse response) {
-        return RestAssured
-                .given().log().all()
+        return given(LINES_GET_BY_ID)
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/lines/{lineId}", response.getId())
                 .then().log().all()
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 지하철_노선_수정_요청(LineResponse response, LineRequest params) {
+    public static ExtractableResponse<Response> 지하철_노선_수정_요청(LineResponse response, LineRequest params, String identifier) {
 
-        return RestAssured
-                .given().log().all()
+        return given(identifier)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(params)
                 .when().put("/lines/" + response.getId())
@@ -76,12 +85,20 @@ public class LineAcceptanceTest extends AcceptanceTest {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 지하철_노선_제거_요청(LineResponse lineResponse) {
-        return RestAssured
-                .given().log().all()
+    public static ExtractableResponse<Response> 지하철_노선_제거_요청(LineResponse lineResponse, String identifier) {
+        return given(identifier)
                 .when().delete("/lines/" + lineResponse.getId())
                 .then().log().all()
                 .extract();
+    }
+
+    public static LineResponse 지하철_노선_등록되어_있음(String name, String color, StationResponse upStation, StationResponse downStation, int distance, Long extraFare) {
+        LineRequest lineRequest = new LineRequest(name, color, upStation.getId(), downStation.getId(), distance, extraFare);
+        return 지하철_노선_등록되어_있음(lineRequest);
+    }
+
+    public static LineResponse 지하철_노선_등록되어_있음(LineRequest lineRequest) {
+        return 지하철_노선_생성_요청(lineRequest, LINES_POST_SUCCESS).as(LineResponse.class);
     }
 
     public static void 지하철_노선_생성됨(ExtractableResponse response) {
@@ -147,23 +164,11 @@ public class LineAcceptanceTest extends AcceptanceTest {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
-    @BeforeEach
-    public void setUp() {
-        super.setUp();
-
-        // given
-        강남역 = 지하철역_등록되어_있음("강남역");
-        광교역 = 지하철역_등록되어_있음("광교역");
-
-        lineRequest1 = new LineRequest("신분당선", "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
-        lineRequest2 = new LineRequest("구신분당선", "bg-red-600", 강남역.getId(), 광교역.getId(), 15);
-    }
-
     @DisplayName("지하철 노선을 생성한다.")
     @Test
     void createLine() {
         // when
-        ExtractableResponse<Response> response = 지하철_노선_생성_요청(lineRequest1);
+        ExtractableResponse<Response> response = 지하철_노선_생성_요청(lineRequest1, LINES_POST_SUCCESS);
 
         // then
         지하철_노선_생성됨(response);
@@ -176,7 +181,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         지하철_노선_등록되어_있음(lineRequest1);
 
         // when
-        ExtractableResponse<Response> response = 지하철_노선_생성_요청(lineRequest1);
+        ExtractableResponse<Response> response = 지하철_노선_생성_요청(lineRequest1, LINES_POST_FAIL_DUPLICATED);
 
         // then
         지하철_노선_생성_실패됨(response);
@@ -190,7 +195,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         LineRequest errorLineRequest = new LineRequest(errorName, "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
 
         //when
-        ExtractableResponse<Response> extractableResponse = 지하철_노선_생성_요청(errorLineRequest);
+        ExtractableResponse<Response> extractableResponse = 지하철_노선_생성_요청(errorLineRequest, LINES_POST_FAIL_NAME_BLANK);
 
         //then
         지하철_노선_생성_실패됨(extractableResponse);
@@ -200,7 +205,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         errorLineRequest = new LineRequest(errorName, "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
 
         //when
-        extractableResponse = 지하철_노선_생성_요청(errorLineRequest);
+        extractableResponse = 지하철_노선_생성_요청(errorLineRequest, DocsIdentifier.LINES_POST_FAIL_NAME_LENGTH);
 
         //then
         지하철_노선_생성_실패됨(extractableResponse);
@@ -210,7 +215,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         errorLineRequest = new LineRequest(errorName, "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
 
         //when
-        extractableResponse = 지하철_노선_생성_요청(errorLineRequest);
+        extractableResponse = 지하철_노선_생성_요청(errorLineRequest, LINES_POST_FAIL_NAME_LENGTH);
 
         //then
         지하철_노선_생성_실패됨(extractableResponse);
@@ -229,6 +234,26 @@ public class LineAcceptanceTest extends AcceptanceTest {
         // then
         지하철_노선_목록_응답됨(response);
         지하철_노선_목록_포함됨(response, Arrays.asList(lineResponse1, lineResponse2));
+    }
+
+    @DisplayName("지하철 전체 노선도를 조회한다.")
+    @Test
+    void getLineMap() {
+        // given
+        LineResponse lineResponse1 = 지하철_노선_등록되어_있음(lineRequest1);
+        LineResponse lineResponse2 = 지하철_노선_등록되어_있음(lineRequest2);
+
+        MapResponse expectedMapResponse1 = new MapResponse(lineResponse1.getId(), lineResponse1.getName(),
+                lineResponse1.getColor(), lineResponse1.getExtraFare(), lineResponse1.getSections());
+
+        MapResponse expectedMapResponse2 = new MapResponse(lineResponse2.getId(), lineResponse2.getName(),
+                lineResponse2.getColor(), lineResponse2.getExtraFare(), lineResponse2.getSections());
+        // when
+        ExtractableResponse<Response> response = 지하철_전체_노선도_조회_요청();
+
+        // then
+        지하철_전체_노선도_응답됨(response);
+        지하철_전체_노선도_포함됨(response, Arrays.asList(expectedMapResponse1, expectedMapResponse2));
     }
 
     @DisplayName("지하철 노선을 조회한다.")
@@ -251,36 +276,10 @@ public class LineAcceptanceTest extends AcceptanceTest {
         LineResponse lineResponse = 지하철_노선_등록되어_있음(lineRequest1);
 
         // when
-        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest2);
+        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest2, LINES_PUT_SUCCESS);
 
         // then
         지하철_노선_수정됨(response);
-    }
-
-    @DisplayName("지하철 노선을 제거한다.")
-    @Test
-    void deleteLine() {
-        // given
-        LineResponse lineResponse = 지하철_노선_등록되어_있음(lineRequest1);
-
-        // when
-        ExtractableResponse<Response> response = 지하철_노선_제거_요청(lineResponse);
-
-        // then
-        지하철_노선_삭제됨(response);
-    }
-
-    @DisplayName("존재하지 않는 지하철 노선을 수정한다.")
-    @Test
-    void updateNotExistLine() throws Exception {
-        //given
-        LineResponse lineResponse = new LineResponse(Long.MAX_VALUE, null, null, null, null);
-
-        //when
-        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest1);
-
-        //then
-        지하철_노선_수정_실패(response);
     }
 
     @DisplayName("중복된 노선으로 수정한다.")
@@ -290,7 +289,20 @@ public class LineAcceptanceTest extends AcceptanceTest {
         LineResponse lineResponse = 지하철_노선_등록되어_있음(lineRequest1);
 
         //when
-        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest1);
+        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest1, LINES_PUT_FAIL_DUPLICATED);
+
+        //then
+        지하철_노선_수정_실패(response);
+    }
+
+    @DisplayName("존재하지 않는 지하철 노선을 수정한다.")
+    @Test
+    void updateNotExistLine() throws Exception {
+        //given
+        LineResponse lineResponse = new LineResponse(Long.MAX_VALUE, null, null, null, null);
+
+        //when
+        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest1, LINES_PUT_FAIL_NOT_EXIST);
 
         //then
         지하철_노선_수정_실패(response);
@@ -308,7 +320,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         LineRequest errorLineRequest = new LineRequest(errorName, "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
 
         //when
-        ExtractableResponse<Response> extractableResponse = 지하철_노선_수정_요청(lineResponse, errorLineRequest);
+        ExtractableResponse<Response> extractableResponse = 지하철_노선_수정_요청(lineResponse, errorLineRequest, LINES_PUT_FAIL_NAME_BLANK);
 
         //then
         지하철_노선_수정_실패(extractableResponse);
@@ -318,7 +330,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         errorLineRequest = new LineRequest(errorName, "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
 
         //when
-        extractableResponse = 지하철_노선_수정_요청(lineResponse, errorLineRequest);
+        extractableResponse = 지하철_노선_수정_요청(lineResponse, errorLineRequest, LINES_PUT_FAIL_NAME_LENGTH);
 
         //then
         지하철_노선_수정_실패(extractableResponse);
@@ -328,15 +340,23 @@ public class LineAcceptanceTest extends AcceptanceTest {
         errorLineRequest = new LineRequest(errorName, "bg-red-600", 강남역.getId(), 광교역.getId(), 10);
 
         //when
-        extractableResponse = 지하철_노선_생성_요청(errorLineRequest);
+        extractableResponse = 지하철_노선_수정_요청(lineResponse, errorLineRequest, LINES_PUT_FAIL_NAME_LENGTH);
 
         //then
-        지하철_노선_생성_실패됨(extractableResponse);
-        //when
-        ExtractableResponse<Response> response = 지하철_노선_수정_요청(lineResponse, lineRequest1);
+        지하철_노선_수정_실패(extractableResponse);
+    }
 
-        //then
-        지하철_노선_수정_실패(response);
+    @DisplayName("지하철 노선을 제거한다.")
+    @Test
+    void deleteLine() {
+        // given
+        LineResponse lineResponse = 지하철_노선_등록되어_있음(lineRequest1);
+
+        // when
+        ExtractableResponse<Response> response = 지하철_노선_제거_요청(lineResponse, LINES_DELETE_SUCCESS);
+
+        // then
+        지하철_노선_삭제됨(response);
     }
 
     @DisplayName("존재하지 않는 지하철 노선을 삭제한다.")
@@ -346,7 +366,7 @@ public class LineAcceptanceTest extends AcceptanceTest {
         LineResponse lineResponse = new LineResponse(Long.MAX_VALUE, null, null, null, null);
 
         //when
-        ExtractableResponse<Response> response = 지하철_노선_제거_요청(lineResponse);
+        ExtractableResponse<Response> response = 지하철_노선_제거_요청(lineResponse, LINES_DELETE_FAIL);
 
         //then
         지하철_노선_삭제_실패(response);
