+++ src/test/java/wooteco/subway/member/MemberAcceptanceTest.java
@@ -1,6 +1,5 @@
 package wooteco.subway.member;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
@@ -13,6 +12,7 @@ import wooteco.subway.member.dto.MemberRequest;
 import wooteco.subway.member.dto.MemberResponse;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static wooteco.subway.DocsIdentifier.*;
 import static wooteco.subway.auth.AuthAcceptanceTest.로그인되어_있음;
 
 public class MemberAcceptanceTest extends AcceptanceTest {
@@ -23,11 +23,10 @@ public class MemberAcceptanceTest extends AcceptanceTest {
     public static final String NEW_PASSWORD = "new_password";
     public static final int NEW_AGE = 30;
 
-    public static ExtractableResponse<Response> 회원_생성을_요청(String email, String password, Integer age) {
+    public static ExtractableResponse<Response> 회원_생성을_요청(String email, String password, Integer age, String identifier) {
         MemberRequest memberRequest = new MemberRequest(email, password, age);
 
-        return RestAssured
-                .given().log().all()
+        return given(identifier)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(memberRequest)
                 .when().post("/members")
@@ -36,8 +35,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
     }
 
     public static ExtractableResponse<Response> 내_회원_정보_조회_요청(TokenResponse tokenResponse) {
-        return RestAssured
-                .given().log().all()
+        return given(MEMBERS_ME_GET)
                 .auth().oauth2(tokenResponse.getAccessToken())
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/members/me")
@@ -46,11 +44,10 @@ public class MemberAcceptanceTest extends AcceptanceTest {
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 내_회원_정보_수정_요청(TokenResponse tokenResponse, String email, String password, Integer age) {
+    public static ExtractableResponse<Response> 내_회원_정보_수정_요청(TokenResponse tokenResponse, String email, String password, Integer age, String identifier) {
         MemberRequest memberRequest = new MemberRequest(email, password, age);
 
-        return RestAssured
-                .given().log().all()
+        return given(identifier)
                 .auth().oauth2(tokenResponse.getAccessToken())
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .body(memberRequest)
@@ -60,8 +57,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
     }
 
     public static ExtractableResponse<Response> 내_회원_삭제_요청(TokenResponse tokenResponse) {
-        return RestAssured
-                .given().log().all()
+        return given(MEMBERS_ME_DELETE)
                 .auth().oauth2(tokenResponse.getAccessToken())
                 .when().delete("/members/me")
                 .then().log().all()
@@ -98,7 +94,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
     @DisplayName("회원 정보를 관리한다.")
     @Test
     void manageMember() {
-        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE);
+        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE, MEMBERS_POST_SUCCESS);
         회원_생성됨(createResponse);
 
         TokenResponse 사용자 = 로그인되어_있음(EMAIL, PASSWORD);
@@ -106,7 +102,7 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> findResponse = 내_회원_정보_조회_요청(사용자);
         회원_정보_조회됨(findResponse, EMAIL, AGE);
 
-        ExtractableResponse<Response> updateResponse = 내_회원_정보_수정_요청(사용자, EMAIL, NEW_PASSWORD, NEW_AGE);
+        ExtractableResponse<Response> updateResponse = 내_회원_정보_수정_요청(사용자, EMAIL, NEW_PASSWORD, NEW_AGE, MEMBERS_ME_PUT_SUCCESS);
         회원_정보_수정됨(updateResponse);
 
         ExtractableResponse<Response> deleteResponse = 내_회원_삭제_요청(사용자);
@@ -120,18 +116,18 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         String overEmail = "abcdefghijklnmopqrstuvwxyz123dsfkldafskldasfkladfskjl456789@google.com";
 
         //when
-        ExtractableResponse<Response> createFailResponse = 회원_생성을_요청(overEmail, PASSWORD, AGE);
+        ExtractableResponse<Response> createFailResponse = 회원_생성을_요청(overEmail, PASSWORD, AGE, MEMBERS_POST_FAIL_EMAIL);
 
         //then
         회원_생성되지않음(createFailResponse);
 
         //given
-        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE);
+        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE, MEMBERS_POST_SUCCESS);
         회원_생성됨(createResponse);
         TokenResponse 사용자 = 로그인되어_있음(EMAIL, PASSWORD);
 
         //when
-        createFailResponse = 내_회원_정보_수정_요청(사용자, overEmail, PASSWORD, AGE);
+        createFailResponse = 내_회원_정보_수정_요청(사용자, overEmail, PASSWORD, AGE, MEMBERS_ME_PUT_FAIL_EMAIL);
 
         //then
         회원_수정되지않음(createFailResponse);
@@ -145,30 +141,30 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         String shortPassword = "abc";
         String longPassword = "abcdefghijklnmopqrstuvwsadsfkasdfdsafjdfsljdsflkasdfjlxyz";
 
-        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE);
+        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE, MEMBERS_POST_SUCCESS);
         회원_생성됨(createResponse);
         TokenResponse 사용자 = 로그인되어_있음(EMAIL, PASSWORD);
 
         //when
-        ExtractableResponse<Response> createFailResponse = 회원_생성을_요청(EMAIL, shortPassword, AGE);
+        ExtractableResponse<Response> createFailResponse = 회원_생성을_요청(EMAIL, shortPassword, AGE, MEMBERS_POST_FAIL_PASSWORD);
 
         //then
         회원_생성되지않음(createFailResponse);
 
         //when
-        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, shortPassword, AGE);
+        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, shortPassword, AGE, MEMBERS_ME_PUT_FAIL_PASSWORD);
 
         //then
         회원_수정되지않음(createFailResponse);
 
         //when
-        createFailResponse = 회원_생성을_요청(EMAIL, longPassword, AGE);
+        createFailResponse = 회원_생성을_요청(EMAIL, longPassword, AGE, MEMBERS_POST_FAIL_PASSWORD);
 
         //then
         회원_생성되지않음(createFailResponse);
 
         //when
-        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, longPassword, AGE);
+        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, longPassword, AGE, MEMBERS_ME_PUT_FAIL_PASSWORD);
 
         //then
         회원_수정되지않음(createFailResponse);
@@ -181,31 +177,31 @@ public class MemberAcceptanceTest extends AcceptanceTest {
         int negativeAge = -1;
         int exceedAge = 201;
 
-        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE);
+        ExtractableResponse<Response> createResponse = 회원_생성을_요청(EMAIL, PASSWORD, AGE, MEMBERS_POST_SUCCESS);
         회원_생성됨(createResponse);
         TokenResponse 사용자 = 로그인되어_있음(EMAIL, PASSWORD);
 
         //when
-        ExtractableResponse<Response> createFailResponse = 회원_생성을_요청(EMAIL, PASSWORD, negativeAge);
+        ExtractableResponse<Response> createFailResponse = 회원_생성을_요청(EMAIL, PASSWORD, negativeAge, MEMBERS_POST_FAIL_AGE);
 
         //then
         회원_생성되지않음(createFailResponse);
 
         //when
-        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, PASSWORD, negativeAge);
+        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, PASSWORD, negativeAge, MEMBERS_ME_PUT_FAIL_AGE);
 
         //then
         회원_수정되지않음(createFailResponse);
 
 
         //when
-        createFailResponse = 회원_생성을_요청(EMAIL, PASSWORD, exceedAge);
+        createFailResponse = 회원_생성을_요청(EMAIL, PASSWORD, exceedAge, MEMBERS_POST_FAIL_AGE);
 
         //then
         회원_생성되지않음(createFailResponse);
 
         //when
-        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, PASSWORD, exceedAge);
+        createFailResponse = 내_회원_정보_수정_요청(사용자, EMAIL, PASSWORD, exceedAge, MEMBERS_ME_PUT_FAIL_AGE);
 
         //then
         회원_수정되지않음(createFailResponse);
