+++ src/test/java/wooteco/subway/station/StationAcceptanceTest.java
@@ -1,6 +1,5 @@
 package wooteco.subway.station;
 
-import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.junit.jupiter.api.DisplayName;
@@ -8,6 +7,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.DocsIdentifier;
 import wooteco.subway.line.dto.LineRequest;
 import wooteco.subway.station.dto.StationRequest;
 import wooteco.subway.station.dto.StationResponse;
@@ -17,6 +17,7 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static wooteco.subway.DocsIdentifier.*;
 import static wooteco.subway.line.LineAcceptanceTest.지하철_노선_등록되어_있음;
 
 @DisplayName("지하철역 관련 기능")
@@ -24,15 +25,10 @@ public class StationAcceptanceTest extends AcceptanceTest {
     private static final String 강남역 = "강남역";
     private static final String 역삼역 = "역삼역";
 
-    public static StationResponse 지하철역_등록되어_있음(String name) {
-        return 지하철역_생성_요청(name).as(StationResponse.class);
-    }
-
-    public static ExtractableResponse<Response> 지하철역_생성_요청(String name) {
+    public static ExtractableResponse<Response> 지하철역_생성_요청(String name, String identifier) {
         StationRequest stationRequest = new StationRequest(name);
 
-        return RestAssured
-                .given().log().all()
+        return given(identifier)
                 .body(stationRequest)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
                 .when().post("/stations")
@@ -41,21 +37,33 @@ public class StationAcceptanceTest extends AcceptanceTest {
     }
 
     public static ExtractableResponse<Response> 지하철역_목록_조회_요청() {
-        return RestAssured
-                .given().log().all()
+        return given(STATIONS_GET)
                 .when().get("/stations")
                 .then().log().all()
                 .extract();
     }
 
-    public static ExtractableResponse<Response> 지하철역_제거_요청(StationResponse stationResponse) {
-        return RestAssured
-                .given().log().all()
+    public static ExtractableResponse<Response> 지하철역_제거_요청(StationResponse stationResponse, String identifier) {
+        return given(identifier)
                 .when().delete("/stations/" + stationResponse.getId())
                 .then().log().all()
                 .extract();
     }
 
+    public static ExtractableResponse<Response> 지하철역_수정_요청(Long id, String name, String identifier) {
+        StationRequest stationRequest = new StationRequest(name);
+        return given(identifier)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(stationRequest)
+                .when().put("/stations/" + id)
+                .then().log().all()
+                .extract();
+    }
+
+    public static StationResponse 지하철역_등록되어_있음(String name) {
+        return 지하철역_생성_요청(name, STATIONS_POST_SUCCESS).as(StationResponse.class);
+    }
+    
     public static void 지하철역_생성됨(ExtractableResponse response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
         assertThat(response.header("Location")).isNotBlank();
@@ -93,16 +101,6 @@ public class StationAcceptanceTest extends AcceptanceTest {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
     }
 
-    public static ExtractableResponse<Response> 지하철역_수정_요청(Long id, String name) {
-        StationRequest stationRequest = new StationRequest(name);
-        return RestAssured
-                .given().log().all()
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .body(stationRequest)
-                .when().put("/stations/" + id)
-                .then().log().all()
-                .extract();
-    }
 
     public static void 지하철역_수정됨(ExtractableResponse response) {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
@@ -116,7 +114,7 @@ public class StationAcceptanceTest extends AcceptanceTest {
     @Test
     void createStation() {
         // when
-        ExtractableResponse<Response> response = 지하철역_생성_요청(강남역);
+        ExtractableResponse<Response> response = 지하철역_생성_요청(강남역, STATIONS_POST_SUCCESS);
 
         // then
         지하철역_생성됨(response);
@@ -126,28 +124,28 @@ public class StationAcceptanceTest extends AcceptanceTest {
     @Test
     void createErrorNameStation() throws Exception {
         //given
-        String errorName = "공백포함 역";
+        String numberName = "공백포함 역";
 
         //when
-        ExtractableResponse<Response> response = 지하철역_생성_요청(errorName);
+        ExtractableResponse<Response> response = 지하철역_생성_요청(numberName, STATIONS_POST_FAIL_NAME_BLANK);
 
         //then
         지하철역_생성되지_않음(response);
 
         //given
-        errorName = "역";
+        numberName = "역";
 
         //when
-        response = 지하철역_생성_요청(errorName);
+        response = 지하철역_생성_요청(numberName, DocsIdentifier.STATIONS_POST_FAIL_NAME_LENGTH);
 
         //then
         지하철역_생성되지_않음(response);
 
         //given
-        errorName = "스무자넘는매우매우매우매우매우매우진짜이상하고정말존재하지않을것같은역";
+        numberName = "스무자넘는매우매우매우매우매우매우진짜이상하고정말존재하지않을것같은역";
 
         //when
-        response = 지하철역_생성_요청(errorName);
+        response = 지하철역_생성_요청(numberName, STATIONS_POST_FAIL_NAME_LENGTH);
 
         //then
         지하철역_생성되지_않음(response);
@@ -160,7 +158,7 @@ public class StationAcceptanceTest extends AcceptanceTest {
         지하철역_등록되어_있음(강남역);
 
         // when
-        ExtractableResponse<Response> response = 지하철역_생성_요청(강남역);
+        ExtractableResponse<Response> response = 지하철역_생성_요청(강남역, STATIONS_POST_FAIL_DUPLICATED);
 
         // then
         지하철역_생성_실패됨(response);
@@ -181,34 +179,6 @@ public class StationAcceptanceTest extends AcceptanceTest {
         지하철역_목록_포함됨(response, Arrays.asList(stationResponse1, stationResponse2));
     }
 
-    @DisplayName("지하철역을 제거한다.")
-    @Test
-    void deleteStation() {
-        // given
-        StationResponse stationResponse = 지하철역_등록되어_있음(강남역);
-
-        // when
-        ExtractableResponse<Response> response = 지하철역_제거_요청(stationResponse);
-
-        // then
-        지하철역_삭제됨(response);
-    }
-
-    @DisplayName("지하철 역 삭제 실패한다. - 이미 구간에 등록된 지하쳘 역")
-    @Test
-    void deleteStation_fail_already_assigned() throws Exception {
-        //given
-        StationResponse stationResponse1 = 지하철역_등록되어_있음(강남역);
-        StationResponse stationResponse2 = 지하철역_등록되어_있음(역삼역);
-        지하철_노선_등록되어_있음(new LineRequest("신분당선", "bg-red-600", stationResponse1.getId(), stationResponse2.getId(), 10));
-
-        //when
-        ExtractableResponse<Response> response = 지하철역_제거_요청(stationResponse1);
-
-        //then
-        지하철역_삭제_실패됨(response);
-    }
-
     @DisplayName("지하철 역 이름을 수정한다.")
     @Test
     void updateStationName_success() throws Exception {
@@ -216,20 +186,20 @@ public class StationAcceptanceTest extends AcceptanceTest {
         StationResponse stationResponse = 지하철역_등록되어_있음(강남역);
 
         //when
-        ExtractableResponse<Response> response = 지하철역_수정_요청(stationResponse.getId(), "송파나루역");
+        ExtractableResponse<Response> response = 지하철역_수정_요청(stationResponse.getId(), "송파나루역", STATIONS_PUT_SUCCESS);
 
         //then
         지하철역_수정됨(response);
     }
 
-    @DisplayName("지하철 역 이름 수정을 실패한다. - 이름 중복")
+    @DisplayName("지하철 역 이름 수정 실패한다. - 이름 중복")
     @Test
     void updateStationName_fail_duplicated() throws Exception {
         //given
         StationResponse stationResponse = 지하철역_등록되어_있음(강남역);
 
         //when
-        ExtractableResponse<Response> response = 지하철역_수정_요청(stationResponse.getId(), "강남역");
+        ExtractableResponse<Response> response = 지하철역_수정_요청(stationResponse.getId(), "강남역", STATIONS_PUT_FAIL_DUPLICATED);
 
         //then
         지하철역_수정_실패(response);
@@ -244,7 +214,7 @@ public class StationAcceptanceTest extends AcceptanceTest {
         String errorName = "공백포함 역";
 
         //when
-        ExtractableResponse<Response> response = 지하철역_수정_요청(stationResponse.getId(), errorName);
+        ExtractableResponse<Response> response = 지하철역_수정_요청(stationResponse.getId(), errorName, STATIONS_PUT_FAIL_NAME_BLANK);
 
         //then
         지하철역_수정_실패(response);
@@ -253,7 +223,7 @@ public class StationAcceptanceTest extends AcceptanceTest {
         errorName = "역";
 
         //when
-        response = 지하철역_수정_요청(stationResponse.getId(), errorName);
+        response = 지하철역_수정_요청(stationResponse.getId(), errorName, STATIONS_PUT_FAIL_NAME_LENGTH);
 
         //then
         지하철역_수정_실패(response);
@@ -262,9 +232,36 @@ public class StationAcceptanceTest extends AcceptanceTest {
         errorName = "스무자넘는매우매우매우매우매우매우진짜이상하고정말존재하지않을것같은역";
 
         //when
-        response = 지하철역_수정_요청(stationResponse.getId(), errorName);
+        response = 지하철역_수정_요청(stationResponse.getId(), errorName, STATIONS_PUT_FAIL_NAME_LENGTH);
 
         //then
         지하철역_수정_실패(response);
+    }
+
+    @DisplayName("지하철역을 제거한다.")
+    @Test
+    void deleteStation() {
+        // given
+        StationResponse stationResponse = 지하철역_등록되어_있음(강남역);
+
+        // when
+        ExtractableResponse<Response> response = 지하철역_제거_요청(stationResponse, STATIONS_DELETE_SUCCESS);
+
+        // then
+        지하철역_삭제됨(response);
+    }
+
+    @DisplayName("지하철 역 삭제 실패한다. - 이미 구간에 등록된 지하쳘 역")
+    @Test
+    void deleteStation_fail_already_assigned() throws Exception {
+        //given
+        StationResponse stationResponse1 = 지하철역_등록되어_있음(강남역);
+        StationResponse stationResponse2 = 지하철역_등록되어_있음(역삼역);
+        지하철_노선_등록되어_있음(new LineRequest("신분당선", "bg-red-600", stationResponse1.getId(), stationResponse2.getId(), 10));
+
+        //when
+        ExtractableResponse<Response> response = 지하철역_제거_요청(stationResponse1, STATIONS_DELETE_FAIL);
+        //then
+        지하철역_삭제_실패됨(response);
     }
 }
